requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic-settings==2.2.1
sqlalchemy==2.0.28
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
boto3==1.34.51
pytest==8.0.2
httpx==0.27.0
email-validator>=2.0.0
python-multipart>=0.0.9
pytest
pytest-mock